[BUGFIX] Tag generated by social viewhelpers hide following tags
The <fb:like> and <fb:comments> tags generated by the view helpers
faceboo.like and facebook.comment are using the XHTML self-closing
syntax, which causes problems in an non - XHTML rendering context
eg. when using HTML5 as doctype.

To fix this, the property forceClosingTag of the tag builder used
is always set to TRUE.

Change-Id: If9772a3f50ace60481918d65a890d3ef78f87d43
Resolves: #44433
Releases: 1.4