[BUGFIX] Better check if current tag has been found
When creating tags with the enhanced suggest wizard from news, you run
into problems with suggestions for partial matches.
Example: add a tag for e.g. "montagmorgen" and then try to add one for
"montag". Since the core suggest function returns a result for the given
string, this will skip the creation of the link to add a new one

Change-Id: I43728ec8d7e50a66783ba2c58b39a2e767d15333
Fixes: #45307