[BUGFIX] logicalOr param must not be an empty array
Make sure the constraints array is not empty before to pass
to the logical disjunction method.

If we want to use the Tx_News_Domain_Model_Dto_Search object
to store additional properties and make usage of the hook available
in generateQuery method of the Abstract repository, the search object
is set but the searchSubject not.
In this case the searchConstraints remain empty and the logicalOr
doesn't like that.

Fixes: #45382
Release: 2.0.1, master
Change-Id: Ib3222902218c336c8b42d740e1f860d9d89ab836