[BUGFIX] Fix of proxy-class generation
ClassCacheBuilder now strips everything outside the class definition
of the extending class to avoid some syntax errors in the generated
file. Before this patch, code outside the class definition of the
extending class would be placed into the class definition of the proxy
class, which would result in syntax errors.

The drawback is that class comments and other potentially useful
things are not moved in the proxy class, but that is a cached class
anyway.

Fixes: #45263
Change-Id: I37b8b4ca6a2ec1ca8216473599e790677d208a18