[BUGFIX] Regenerate proxy classes after cache-clearing
Proxy classes were only generated in the ext_autoload.php file, which
is executed only when installing an extension (as far as I can tell).
This had the consequence that proxy classes were deleted on cache
clearing, but were only regenerated on installation of an extension.

To resolve this, the generation mechanism for proxy classes is called
in the 'clearCachePostProc'-hook.

Related: #45263
Change-Id: I8ec8026a7b2ccae0633e05d0bb4e054ae9eeb70f
Reviewed-on: https://review.typo3.org/18217
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer