[BUGFIX] Make preview of hidden records work in single view
The preview of hidden records currently only works
when a specific news record is selected in the plugin.

This is the case because the property mapper of extbase
already tries to fetch the record for the controller
argument.

To work around this problem, now a different argument is
used to pass the news uid to the action and adapt the
TCEMain hook to use this argument name.

Fixes: #47699
Change-Id: I6cb538387c84ce396a31a7663cf47c60a3df591c