[BUGFIX] Make enabling preview of hidden records in TS work
Because Extbase merges TS and FF configuration by
giving the flexform precedence, it is never possible
to set the value by TypoScript because the FlexForm
field has the same name as in TypoScript.

Solution is to make the field in the Flexform a selectbox
and use another name which is checked in TS.

This is particulary useful for enabling this setting
using a condition for logged in BE-users

As this never worked, we can just rename the TypoScript
setting name without breaking functionality.

Fixes: #47704
Change-Id: I7af7760fd6bc8c3cdcfd3c3839984d29a94f5428