[FEATURE] Improved recursive setting inheritance
The recursive level of the starting point can be reset to zero in the
FlexForm settings but can also set to an empty string so that the
setting is inherited from the TypoScript settings.

For this to work the overrideFlexformSettingsIfEmpty behavior is
modified, so that it only will treat empty strings as empty values but
zero integers will be treated as set values.

Change-Id: I5c3f003c000d99ec4e3a675e312468aa3909ed48
Resolves: #54090
Reviewed-on: https://review.typo3.org/25780
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer