[!!!][TASK] Use sys_category
As of TYPO3 6.0 sys_categories are part of the core and
since TYPO3 6.2 is has the same BeUsers category mount options
as we known from ext:news.

This patch extends the sys_categorie record with the fields
we know from tx_news_domain_model_category. And changes ext:news
to use sys_categorie instead.

These fields are added to sys_category:
- fe_group
- single_pid
- shortcut
- import_id
- import_source

Make sure that the new sys_category fields are added before
running the updater in the extension manager after appling
this patch. In normal extension update procedure this is
done automaticly by extension manager.

Resolves: #47020
Change-Id: I5a7faa70b157e1ea1c846f2fbdac1c454462a1b7
Reviewed-on: https://review.typo3.org/28163
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer