[BUGFIX] Localized categories must point to parents after import
When importing native categories to sys_category, the l10n_parent
of categories having an l10n_parent (= localized categories with
default translation) must point to the new parents.

To achieve this, the import is splitted in an import for default
language categories that builds a mapping array. Then the localized
categories are imported. If a mapping is found, the l10n_parent is
set to the imported parent category.

Resolves: #59201
Releases: 3.0
Change-Id: I70571b521944945603c0fe3b2c0fdcb0d731aa20
Reviewed-on: https://review.typo3.org/30775
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer