[BUGFIX] Add switch for category restriction check
Feature ##59063 introduces a forced check on category permissions.

This patch introduces a switch for this behaviour so you can
disable this and a editor can save the news items just as before
if he hasn't access to all selected categories.

By default the check is disabled as this was the old default behaviour.

Resolves: #61068

Change-Id: I0109a779e67ffb605dd4d1fa0cb6cb71fd837dd5
Reviewed-on: http://review.typo3.org/32244
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>