[BUGFIX] Class Cache Manager must work with CRLF
The Class Cache Manager generated a Cache file based on the
content of a file. The generated file contains an faulty
opening curly brace if the extending class is formatted
in PSR-2 (opening curly brace of class is in a new line) and
if the file contains CRLF (Windows linebreaks) instead of LF.

CRLF's must be replaced to LF before the content of the file
is parsed.

Change-Id: Ia3178f8b88ab7d0dc106b9d53a11e3ae2a0b2e29
Resolves: #71560
Releases: master
Reviewed-on: https://review.typo3.org/44705
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>