[BUGFIX] Preview of hidden news not working for translated news
The preview of hidden news records doesn't work for translated news
records. Assuming language mode strict, this is the situation:

* Hidden news in default language: OK
* Hidden translation of a hidden news: Page not found
* Hidden news that only exists in a non-default language: OK
* News with language [All]: OK

The problem can be solved by enabling PageRepository::showHiddenRecords
and using the original uid of the news, not the uid of its l10n_parent,
for the preview of translated news.

Resolves: #72363
Releases: master, 3.2
Change-Id: Ifc4bfe66feb704bfef3638526d7ab4886524a457
Reviewed-on: https://review.typo3.org/46778
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>