[TASK] Adapt to breaking core hook getFlexFormDSClass
Flex form manipulation via hook getFlexFormDSClass in method
getFlexFormDS() will no longer work with core version >= 8.5
since the method is not called anymore within core.

The current usage of the hook basically hides fields flex form
fields in FormEngine and adds a foreign_table_where in another
field.

This does not affect the data handler, so this hook can be
transferred to a FormEngine data provider that is placed
between TcaFlexPrepare and TcaFlexProcess, not using the
additionally available hooks of the new core data structure
handling.

Related: #78581