[!!!][BUGFIX] Change slug evaluation to unique instead of uniqueInSite
The slug eval configuration of all 3 slug fields (news,tag and category
record) is changed from `uniqueInSite` to `unique`.

If you really need `uniqueInSite` you can switch to this setting in the
Extension settings in the Install Tool.

Some further explanation:
If `uniqueInSite` is used, news records in multiple sites can have the
slug. The drawback is that those news can't be shared between these
sites.
Using `unique` means that news records in multiple sites have always a
different slug, no matter if the title is the same or not. Therefore
news can be shared between multiple sites.