[DOCS] Changes related to Extension Configuration (#1672)
The extension is no longer configured in the Extension Manager,
but configured in the Extension Configuration. Some files, labels
and decriptions are changed to reflect that.

Also in some internal references (:ref:) the anchor text was removed.
This way the header of the linked to section will automatically be
used as anchor text in the link.