[DOCS] Add portal-like startpages (#1679)
* Working on the startpage

* [TASK] move files to more logic positions

releases: master

* Remove documentation examples, move missing errors

* Create Quick start

* [BUGFIX] Properties of type DateTimeInterface should be nullable

Resolves: #1665

* [DOCS] Fix formatting for typoscript (#1674)

The textrole tsconfig does not exist. Since it is TypoScript
syntax, the role typoscript can be used, as in:

:typoscript:`something`.

See Includes.rst.txt

* [DOCS] Improve style (#1675)

Fix double word.

Also, using terms like "easy" etc. to describe something is discouraged,
see Symfony docs styleguide.

* [DOCS] Changes related to Extension Configuration (#1672)

The extension is no longer configured in the Extension Manager,
but configured in the Extension Configuration. Some files, labels
and decriptions are changed to reflect that.

Also in some internal references (:ref:) the anchor text was removed.
This way the header of the linked to section will automatically be
used as anchor text in the link.

* [DOCS] Minor improvements on thanks page (#1669)

* [DOCS] Minor improvements on How to start page (#1670)

* [TASK] move files to more logic positions

releases: master

* Remove documentation examples, move missing errors

* Create Quick start

* Working on the startpage

* Move files to logic posititions, create portal pages

* Fix rendering warnings, make an overview of tutorials

* Remove utf-8 -*- with BOM information

Update User manual, add information on using the News admin module and the list module

Co-authored-by: lina.wolf <lwolf@w-commerce.de>
Co-authored-by: Georg Ringer <georg.ringer@gmail.com>
Co-authored-by: Sybille Peters <sypets@gmx.de>