[DOCS] Align with new TYPO3 documentation standards (#1771)
* [DOCS] Align with new TYPO3 documentation standards

- align README.md, Index.rst, Includes.rst.txt, Settings.cfg
- add genindex.rst, Sitemap.rst
- remove outdated Targets.rst

See https://docs.typo3.org/m/typo3/docs-how-to-document/main/en-us/GeneralConventions/FileStructure.html for further details.

* [DOCS] Reference manual's start page with :doc:`<manual>:Index`

Adding the custom label `start` to the beginning of the manual's
Index.rst is redundant. Use :doc:`<manual>:Index` instead of
:ref:`<manual:start>` to refer to it.

* [TASK] Add TER + documentation + repository to composer.json

Add the common extension destinations to composer.json. Packagist
displays them in a prominent place.

* [DOCS] Replace :ts: with :typoscript: text role

The ambiguous :ts: text role has been removed to
not confuse the writer with typescript and typoscript.

* [DOCS] Fix rendering warnings

* [DOCS] Aggregate and update the installation guide

- move the installation section of README.md into
  Documentation/Administration/Installation/Index.rst
- move the configuration section from the Installation
  page to an own Configuration page
- add extension compatibility matrix into Installation page

* [DOCS] Aggregate and update the quick start guide

Move the quick start section of README.md into Documentation/.

* [DOCS] Aggregate and update the support guide

Move the sponsoring section of README.md into Documentation/
and link to it.

* [DOCS] Aggregate and update the administration guide

Move the administration section of README.md into Documentation/.

* [DOCS] Link the contribution guide in README.md

* [DOCS] Remove unnecessary "About" page

All information on this page is already included
elsewhere in the documentation.

* [DOCS] Remove superfluous `only:: html`

This directive is used for hiding TOCs in printed editions
of the documentation, but should not hide text.

* [DOCS] Fix quotation in Introduction/About page

* [DOCS] Fix english grammar

* [DOCS] Rework Contribute, Support and Sponsoring pages

- rename "Contribute" to "Contribution" to align with "Sponsoring"
- rename "Thanks" to "Sponsoring" to be more precise
- aggregate and update "Sponsoring" page
- reword small parts of "Contribution" page

* [DOCS] Adapt quotation in README.md