[FEATURE] Introduce simplified AliasMapper for News Routing
Route Enhancer Configuration can be made easier for
EXT:news.

This change adds a new Route Enhancer (NewsAliasMapper),
which is a shorthand for the PersistedAliasMapper, it is
registered with three custom types (news, categories + tags).

Before in your site configuration:

       aspects:
         news-title:
           type: PersistedAliasMapper
           tableName: tx_news_domain_model_news
           routeFieldName: path_segment
           type: NewsTitle

Now:

       aspects:
         news-title:
           type: NewsTitle

The three new types are "NewsTitle" (for news records),
"NewsTag" (for tag records) and "NewsCategory" (for sys_category
records).

In addition, it supports the possibility that if no
record (news, tag or category) was found, "0" is returned,
thus having the News plugin to handle a invalid news identifier
(e.g. a news which is now hidden).