[TASK] Use current standard for web-dir
To make contribution and maintainance of configuration, scripts etc.
for testing easier, we try to use a common standard across TYPO3
extensions.

The current standard for web-dir is .Build/Web and is used in the
official example extensions (tea, docs-examples) as well as in
most of the checked community extensions (in a subset of 20 of
the most popular v12 extensions).

Related: TYPO3-Documentation/tea#802