[DOCS] Modernize outdated conditions that still use globalVar
Since TYPO3 v10 the globalVar notation is deprecated. This patch converts all conditions to the new syntax.

A null-safe operator was deliberately not used in this patch, as this was only introduced from TYPO3 v12 onwards. As soon as support for v11 has ended, the shorthand `[getTSFE()?.id == 9819]` can be used accordingly.

In addition, `[global]` is replaced by the more suitable variant `[END]`, which only terminates the previous condition.