[DOCS] Modernize outdated conditions still using globalVar
With the introduction of the symfony expression language since TYPO3 9.4, the old syntax (such as the use of `globalVar`) is considered as deprecated. This patch converts all conditions to the new syntax.

A [null-safe operator in the TypoScript conditions](https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.1/Feature-100586-Null-safeOperatorInTypoScriptConditions.html#feature-100586-null-safe-operator-in-typoscript-conditions) was deliberately not used in this patch, as this was only introduced from TYPO3 v12.1 onwards. As soon as support for v11 is ended, this more elegant abbreviation `[getTSFE()?.type == 9819]` can be used accordingly.

In addition, `[global]` is replaced by the more suitable variant `[END]`, which only terminates the previous condition.

Related: #2458